requirements.txt
appnope==0.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.5
djangorestframework==3.6.4
idna==2.6
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.10.2
ldap3==2.3
osmapi==1.0.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.4
Pygments==2.2.0
pyldap==2.4.37
pytz==2017.2
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
